+++ lib/nes/cartridge/mapper/mapper.dart
@@ -5,6 +5,7 @@ import 'package:nesd/nes/cartridge/cartridge.dart';
 import 'package:nesd/nes/cartridge/mapper/axrom.dart';
 import 'package:nesd/nes/cartridge/mapper/br909x.dart';
 import 'package:nesd/nes/cartridge/mapper/cnrom.dart';
+import 'package:nesd/nes/cartridge/mapper/gxrom.dart';
 import 'package:nesd/nes/cartridge/mapper/mapper_state.dart';
 import 'package:nesd/nes/cartridge/mapper/mmc1.dart';
 import 'package:nesd/nes/cartridge/mapper/mmc2.dart';
@@ -25,6 +26,7 @@ abstract class Mapper {
       4 => MMC3(),
       7 => AxROM(),
       9 => MMC2(),
+      66 => GxROM(),
       71 => BR909x(),
       206 => Namco108(),
       _ => throw UnsupportedMapper(mapper),